Restructured folders and reset Jest setup
