+++ frontend/package.json
@@ -3,13 +3,14 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@testing-library/jest-dom": "^5.16.5",
-		"@testing-library/react": "^14.0.0",
+		"@testing-library/jest-dom": "^5.14.1",
+		"@testing-library/react": "^13.0.0",
 		"@testing-library/user-event": "^13.2.1",
-		"@types/jest": "^29.5.1",
+		"@types/jest": "^27.0.1",
 		"@types/node": "^16.7.13",
 		"@types/react": "^18.0.0",
 		"@types/react-dom": "^18.0.0",
+		"@types/react-test-renderer": "^18.0.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-scripts": "5.0.1",
@@ -19,7 +20,7 @@
 	"scripts": {
 		"start": "react-scripts start",
 		"build": "react-scripts build",
-		"test": "jest",
+		"test": "react-scripts test",
 		"eject": "react-scripts eject",
 		"lint": "eslint .",
 		"lint:fix": "eslint --fix",
@@ -44,22 +45,13 @@
 		]
 	},
 	"devDependencies": {
-		"@babel/preset-env": "^7.21.4",
-		"@babel/preset-react": "^7.18.6",
-		"@babel/preset-typescript": "^7.21.4",
-		"@jest/globals": "^29.5.0",
-		"@testing-library/react": "^14.0.0",
 		"@typescript-eslint/eslint-plugin": "^5.59.1",
 		"@typescript-eslint/parser": "^5.59.1",
-		"babel-jest": "^29.5.0",
 		"eslint": "^8.39.0",
 		"eslint-config-prettier": "^8.8.0",
 		"eslint-plugin-airbnb": "^0.0.1-security",
 		"eslint-plugin-prettier": "^4.2.1",
 		"eslint-plugin-react": "^7.32.2",
-		"jest": "^29.5.0",
-		"prettier": "^2.8.8",
-		"react-test-renderer": "^18.2.0",
-		"ts-jest": "^29.1.0"
+		"prettier": "^2.8.8"
 	}
 }
